tests: rename `test` package to `tests`
