+++ .github/workflows/running-tests.yaml
@@ -23,4 +23,4 @@ jobs:
         run: pip install PyYAML
 
       - name: Run unit tests
-        run: python kate/test/capabilities_test.py
+        run: python kate/tests/capabilities_test.py
