+++ kate/tests/__init__.py
