+++ kate/tests/helper.py
